move immatriculation components to own folder in common
