+++ src/components/modals/ImmatriculationHistoryModal.vue
@@ -21,7 +21,7 @@
 <script lang="ts">
     import Modal from "@/components/modals/Modal.vue";
     import { ref, onBeforeMount } from "vue";
-    import ImmatriculationHistory from "@/components/ImmatriculationHistory.vue";
+    import ImmatriculationHistory from "@/components/common/immatriculation/ImmatriculationHistory.vue";
     import LoadingSpinner from "@/components/loading/Spinner.vue";
     import { useStore } from "@/store/store";
